added support for unilateral movement by changing the type of FaceRequest message types.
The au is now sent as a string, that can optionally have an 'l' or an 'r'
appended to the end of the au, indicating which side should be moved.
Appending no letter to the end automatically moves both sides, and any
au can be sent unilaterally, not just eyebrows.